refactor(background): extract X session auth/transport from fast-batch
The webRequest auth capture, session-storage persistence, permission
request, authed GraphQL fetch, and error classification (~130 lines) move
out of fast-batch.ts into x-session.ts. fast-batch now imports that
surface and reads the live `templates` map; its orchestration (pagination,
expansion, write pipeline, messages) stays put. Strict move — logic and
console output unchanged; typecheck, build, and 194 tests green.
